Remove "docker-compose" alias from docker:latest
Close #33 #32 
The latest build introduced a breaking change and forced upon users the compose CLI v2 that embraced the newer compose-spec.
This patch restores the "docker-compose" to the expected v1.

Thus, users are encouraged to migrate whenever possible to v2. Beware that the ecosystem may not be ready depending on what you use, e.g. `docker stack` command does not support the newer specification.